+++ navground_core/include/navground/core/kinematics.h
@@ -1,16 +1,16 @@
 #ifndef NAVGROUND_CORE_KINEMATICS_H_
 #define NAVGROUND_CORE_KINEMATICS_H_
 
-#include <assert.h>
-
 #include <algorithm>
+#include <assert.h>
+#include <limits>
 #include <vector>
 
 #include "navground/core/common.h"
+#include "navground/core/export.h"
 #include "navground/core/property.h"
 #include "navground/core/register.h"
 #include "navground/core/types.h"
-#include "navground/core/export.h"
 
 namespace navground::core {
 
@@ -23,18 +23,25 @@ namespace navground::core {
  *
  * - convert between wheel speeds and body twist
  *
- * - store maximal linear and angular speed
+ * - returns maximal linear and angular speed
  *
- * - store the number of degrees of freedom
+ * - returns the number of degrees of freedom
+ *
+ * Negative speed means unconstrained.
  */
 class NAVGROUND_CORE_EXPORT Kinematics
     : virtual public HasProperties,
       virtual public HasRegister<Kinematics> {
- public:
+public:
   using HasRegister<Kinematics>::C;
 
-  Kinematics(ng_float_t max_speed, float max_angular_speed = 0)
-      : max_speed(max_speed), max_angular_speed(max_angular_speed) {}
+  static constexpr ng_float_t inf = std::numeric_limits<ng_float_t>::infinity();
+
+  Kinematics(ng_float_t max_speed = Kinematics::inf,
+             float max_angular_speed = Kinematics::inf)
+      : _max_speed(max_speed < 0 ? Kinematics::inf : max_speed),
+        _max_angular_speed(max_angular_speed < 0 ? Kinematics::inf
+                                                 : max_angular_speed) {}
 
   virtual ~Kinematics() = default;
 
@@ -60,7 +67,7 @@ class NAVGROUND_CORE_EXPORT Kinematics
    * @return     The same desired twist if feasible else the nearest feasible
    * value. How this is defined depends on the concrete sub-class.
    */
-  virtual Twist2 feasible(const Twist2& twist) const = 0;
+  virtual Twist2 feasible(const Twist2 &twist) const = 0;
 
   /**
    * @brief      Computes the nearest feasible twist to a desired twist,
@@ -73,8 +80,8 @@ class NAVGROUND_CORE_EXPORT Kinematics
    * @return     The same desired twist if feasible else the nearest feasible
    * value. How this is defined depends on the concrete sub-class.
    */
-  virtual Twist2 feasible(const Twist2& twist,
-                          [[maybe_unused]] const Twist2& current,
+  virtual Twist2 feasible(const Twist2 &twist,
+                          [[maybe_unused]] const Twist2 &current,
                           [[maybe_unused]] ng_float_t time_step) const {
     return feasible(twist);
   }
@@ -97,60 +104,72 @@ class NAVGROUND_CORE_EXPORT Kinematics
    *
    * @return     The maximal speed.
    */
-  ng_float_t get_max_speed() const { return max_speed; }
+  ng_float_t get_max_speed() const { return _max_speed; }
   /**
    * @brief      Sets the maximum speed.
    *
-   * @param[in]  value  A positive value.
+   * @param[in]  value  The desired value.
+   *                    A negative number is interpreted as +infinite.
    */
   void set_max_speed(ng_float_t value) {
-    max_speed = std::max<ng_float_t>(0, value);
+    if (value < 0) {
+      _max_speed = Kinematics::inf;
+    } else {
+      _max_speed = value;
+    }
   }
   /**
    * @brief      Gets the maximal angular speed.
    *
    * @return     The maximal angular speed.
    */
-  virtual ng_float_t get_max_angular_speed() const { return max_angular_speed; }
+  virtual ng_float_t get_max_angular_speed() const {
+    return _max_angular_speed;
+  }
   /**
    * @brief      Sets the maximum angular speed.
    *
-   * @param[in]  value  A positive value.
+   * @param[in]  value  The desired value.
+   *                    A negative number is interpreted as +infinite.
    */
   void set_max_angular_speed(ng_float_t value) {
-    max_angular_speed = std::max<ng_float_t>(0, value);
+    if (value < 0) {
+      _max_angular_speed = Kinematics::inf;
+    } else {
+      _max_angular_speed = value;
+    }
   }
 
- private:
+private:
   /**
    * The maximal speed
    */
-  ng_float_t max_speed;
+  ng_float_t _max_speed;
   /**
    * The maximal angular speed
    */
-  ng_float_t max_angular_speed;
+  ng_float_t _max_angular_speed;
 };
 
 /**
  * @brief      Unconstrained kinematics (e.g., quad-copters)
  */
 class NAVGROUND_CORE_EXPORT OmnidirectionalKinematics : public Kinematics {
- public:
+public:
   /**
    * @brief      Constructs a new instance.
    *
    * @param[in]  max_speed          The maximal speed
    * @param[in]  max_angular_speed  The maximal angular speed
    */
-  OmnidirectionalKinematics(ng_float_t max_speed = 0,
-                            ng_float_t max_angular_speed = 0)
+  OmnidirectionalKinematics(ng_float_t max_speed = Kinematics::inf,
+                            ng_float_t max_angular_speed = Kinematics::inf)
       : Kinematics(max_speed, max_angular_speed) {}
 
   /**
    * @private
    */
-  Twist2 feasible(const Twist2& twist) const override;
+  Twist2 feasible(const Twist2 &twist) const override;
   /**
    * @brief      Returns whether the kinematics has wheels.
    *
@@ -169,7 +188,7 @@ class NAVGROUND_CORE_EXPORT OmnidirectionalKinematics : public Kinematics {
    */
   std::string get_type() const override { return type; }
 
- private:
+private:
   const static std::string type;
 };
 
@@ -178,20 +197,21 @@ class NAVGROUND_CORE_EXPORT OmnidirectionalKinematics : public Kinematics {
  * move (e.g., people)
  */
 class NAVGROUND_CORE_EXPORT AheadKinematics : public Kinematics {
- public:
+public:
   /**
    * @brief      Constructs a new instance.
    *
    * @param[in]  max_speed          The maximal speed
    * @param[in]  max_angular_speed  The maximal angular speed
    */
-  AheadKinematics(ng_float_t max_speed = 0, ng_float_t max_angular_speed = 0)
+  AheadKinematics(ng_float_t max_speed = Kinematics::inf,
+                  ng_float_t max_angular_speed = Kinematics::inf)
       : Kinematics(max_speed, max_angular_speed) {}
 
   /**
    * @private
    */
-  Twist2 feasible(const Twist2& twist) const override;
+  Twist2 feasible(const Twist2 &twist) const override;
   /**
    * @brief      Returns whether the kinematics has wheels.
    *
@@ -210,7 +230,7 @@ class NAVGROUND_CORE_EXPORT AheadKinematics : public Kinematics {
    */
   std::string get_type() const override { return type; }
 
- private:
+private:
   static const std::string type;
 };
 
@@ -222,9 +242,10 @@ class NAVGROUND_CORE_EXPORT AheadKinematics : public Kinematics {
  *   - `wheel_axis` (float, \ref get_axis)
  */
 class NAVGROUND_CORE_EXPORT WheeledKinematics : public Kinematics {
- public:
-  WheeledKinematics(ng_float_t max_speed, ng_float_t max_angular_speed,
-                    ng_float_t axis)
+public:
+  WheeledKinematics(ng_float_t max_speed = Kinematics::inf,
+                    ng_float_t max_angular_speed = Kinematics::inf,
+                    ng_float_t axis = 0)
       : Kinematics(max_speed, max_angular_speed), axis(axis) {}
 
   virtual ~WheeledKinematics() = default;
@@ -243,7 +264,7 @@ class NAVGROUND_CORE_EXPORT WheeledKinematics : public Kinematics {
    *
    * @return     The corresponding twist
    */
-  virtual Twist2 twist(const WheelSpeeds& value) const = 0;
+  virtual Twist2 twist(const WheelSpeeds &value) const = 0;
   /**
    * @brief      Convert a twist to wheel speeds
    *
@@ -251,7 +272,7 @@ class NAVGROUND_CORE_EXPORT WheeledKinematics : public Kinematics {
    *
    * @return     The corresponding wheel speeds.
    */
-  virtual WheelSpeeds wheel_speeds(const Twist2& value) const = 0;
+  virtual WheelSpeeds wheel_speeds(const Twist2 &value) const = 0;
 
   /**
    * @brief      Convert a twist to feasible wheel speeds
@@ -260,12 +281,12 @@ class NAVGROUND_CORE_EXPORT WheeledKinematics : public Kinematics {
    *
    * @return     The corresponding wheel speeds.
    */
-  virtual WheelSpeeds feasible_wheel_speeds(const Twist2& value) const = 0;
+  virtual WheelSpeeds feasible_wheel_speeds(const Twist2 &value) const = 0;
 
   /**
    * @private
    */
-  Twist2 feasible(const Twist2& value) const override;
+  Twist2 feasible(const Twist2 &value) const override;
 
   /**
    * @brief      Gets the wheel axis.
@@ -279,13 +300,14 @@ class NAVGROUND_CORE_EXPORT WheeledKinematics : public Kinematics {
    * @param[in]  value  A positive value
    */
   void set_axis(ng_float_t value) {
-    if (value > 0) axis = value;
+    if (value > 0)
+      axis = value;
   }
 
   /**
    * @private
    */
-  virtual const Properties& get_properties() const override {
+  virtual const Properties &get_properties() const override {
     return properties;
   };
 
@@ -294,27 +316,52 @@ class NAVGROUND_CORE_EXPORT WheeledKinematics : public Kinematics {
    */
   static const std::map<std::string, Property> properties;
 
- protected:
+protected:
   ng_float_t axis;
 };
 
 /**
  * @brief   Two differential drive wheels (left, right) (e.g., a wheelchair)
+ *          with software-limited speeds.
+ *
+ * It implements two types of kinematics constrains:
+ *
+ * - a maximal wheel speed that adds a linear coupling between linear and
+ * angular speed (e.g., the robot can move at maximal linear speed only
+ * straights) (same as \ref TwoWheelsDifferentialDriveKinematics)
+ * - a controller that clamps linear and angular speed inside a box: these
+ * constrains when the related bounds are set to negative values.
+ *
+ * *Registered properties*:
+ *
+ *   - `max_forward_speed` (float, \ref get_max_forward_speed)
+ *   - `max_backward_speed` (float, \ref get_max_backward_speed)
+ *
  */
 class NAVGROUND_CORE_EXPORT TwoWheelsDifferentialDriveKinematics
     : public WheeledKinematics {
- public:
+public:
   /**
    * @brief      Constructs a new instance.
    *
    * @param[in]  max_speed          The maximal wheel speed
    * @param[in]  axis               The wheel axis (i.e., the distance between
    * the wheels)
+   * @param[in]  max_angular_speed  The maximal angular speed, that is also
+   * limited by the 2 * max_speed / axis, if the axis is positive.
+   * @param[in]  max_forward_speed  The maximal linear speed when moving
+   * forwards (set to negative or infinite to leave unconstrained)
+   * @param[in]  max_backward_speed The maximal linear speed when moving
+   * backwards (set to negative or infinite to leave unconstrained)
    */
-  TwoWheelsDifferentialDriveKinematics(ng_float_t max_speed = 0,
-                                       ng_float_t axis = 0)
-      : WheeledKinematics(max_speed, (axis > 0) ? 2 * max_speed / axis : 0,
-                          axis) {}
+  TwoWheelsDifferentialDriveKinematics(
+      ng_float_t max_speed = Kinematics::inf, ng_float_t axis = 0,
+      ng_float_t max_angular_speed = Kinematics::inf,
+      ng_float_t max_forward_speed = Kinematics::inf,
+      ng_float_t max_backward_speed = 0)
+      : WheeledKinematics(max_speed, max_angular_speed, axis),
+        _max_forward_speed(max_forward_speed),
+        _max_backward_speed(max_backward_speed) {}
 
   /**
    * @brief      Returns the degrees of freedom
@@ -327,10 +374,67 @@ class NAVGROUND_CORE_EXPORT TwoWheelsDifferentialDriveKinematics
    * @private
    */
   ng_float_t get_max_angular_speed() const override {
+    ng_float_t value = Kinematics::inf;
     if (get_axis() > 0) {
-      return 2 * get_max_speed() / get_axis();
+      value = 2 * get_max_speed() / get_axis();
     }
-    return 0;
+    return std::min(value, Kinematics::get_max_angular_speed());
+  }
+  /**
+   * @brief      Whether the agent can move backwards.
+   *
+   * @return     True if moving backwards is possible.
+   */
+  bool can_move_backwards() const {
+    return get_max_speed() > 0 && _max_backward_speed > 0;
+  }
+  /**
+   * @brief      Whether the agent can move forwards.
+   *
+   * @param[in]  True if moving forwards is possible.
+   */
+  bool can_move_forwards() const {
+    return get_max_speed() > 0 && _max_forward_speed > 0;
+  }
+
+  /**
+   * @brief      Gets the maximal linear speed when moving forwards.
+   *
+   * @return     The maximal speed.
+   */
+  ng_float_t get_max_forward_speed() const { return _max_forward_speed; }
+  /**
+   * @brief      Sets the maximal speed when moving forwards.
+   *
+   * @param[in]  value  The desired value.
+   *                    A negative number is interpreted as +infinite.
+   */
+  void set_max_forward_speed(ng_float_t value) {
+    if (value < 0) {
+      _max_forward_speed = Kinematics::inf;
+    } else {
+      _max_forward_speed = value;
+    }
+  }
+
+  /**
+   * @brief      Gets the maximal linear speed when moving backwards.
+   *
+   * @return     The maximal speed.
+   */
+  ng_float_t get_max_backward_speed() const { return _max_backward_speed; }
+  /**
+   * @brief      Sets the maximal speed when moving backwards.
+   *
+   * @param[in]  value  The desired value.
+   *                    A negative number is interpreted as +infinite.
+   */
+  void set_max_backward_speed(ng_float_t value) {
+    if (value < 0) {
+      _max_backward_speed = Kinematics::inf;
+    } else {
+      _max_backward_speed = value;
+    }
   }
 
   /**
@@ -340,7 +444,7 @@ class NAVGROUND_CORE_EXPORT TwoWheelsDifferentialDriveKinematics
    *
    * @return     The corresponding twist
    */
-  Twist2 twist(const WheelSpeeds& speeds) const override;
+  Twist2 twist(const WheelSpeeds &speeds) const override;
 
   /**
    * @brief      See \ref WheeledKinematics::wheel_speeds.
@@ -349,7 +453,7 @@ class NAVGROUND_CORE_EXPORT TwoWheelsDifferentialDriveKinematics
    *
    * @return     The corresponding wheel speeds in the order {left, right}
    */
-  WheelSpeeds wheel_speeds(const Twist2& twist) const override;
+  WheelSpeeds wheel_speeds(const Twist2 &twist) const override;
 
   /**
    * @brief      See \ref WheeledKinematics::feasible_wheel_speeds.
@@ -358,19 +462,31 @@ class NAVGROUND_CORE_EXPORT TwoWheelsDifferentialDriveKinematics
    *
    * @return     The corresponding wheel speeds in the order {left, right}
    */
-  WheelSpeeds feasible_wheel_speeds(const Twist2& twist) const override;
+  WheelSpeeds feasible_wheel_speeds(const Twist2 &twist) const override;
 
   /** @private
    */
-  Twist2 feasible(const Twist2& twist) const override;
+  Twist2 feasible(const Twist2 &twist) const override;
 
   /**
    * @private
    */
   std::string get_type() const override { return type; }
 
- private:
+  /**
+   * @private
+   */
+  const Properties &get_properties() const override { return properties; };
+
+  /**
+   * @private
+   */
+  static const std::map<std::string, Property> properties;
+
+private:
   const static std::string type;
+  ng_float_t _max_forward_speed;
+  ng_float_t _max_backward_speed;
 };
 
 // TODO(Jerome): make it general
@@ -383,7 +499,7 @@ class NAVGROUND_CORE_EXPORT TwoWheelsDifferentialDriveKinematics
  */
 class NAVGROUND_CORE_EXPORT FourWheelsOmniDriveKinematics
     : public WheeledKinematics {
- public:
+public:
   /**
    * @brief      Constructs a new instance.
    *
@@ -391,8 +507,11 @@ class NAVGROUND_CORE_EXPORT FourWheelsOmniDriveKinematics
    * @param[in]  axis               The wheel axis (i.e., the distance between
    * the wheels)
    */
-  FourWheelsOmniDriveKinematics(ng_float_t max_speed = 0, ng_float_t axis = 0)
-      : WheeledKinematics(max_speed, axis > 0 ? max_speed / axis : 0, axis) {}
+  FourWheelsOmniDriveKinematics(ng_float_t max_speed = Kinematics::inf,
+                                ng_float_t axis = 0)
+      : WheeledKinematics(max_speed,
+                          axis > 0 ? max_speed / axis : Kinematics::inf, axis) {
+  }
 
   /**
    * @brief      Returns the degrees of freedom
@@ -419,7 +538,7 @@ class NAVGROUND_CORE_EXPORT FourWheelsOmniDriveKinematics
    *
    * @return     The corresponding twist
    */
-  Twist2 twist(const WheelSpeeds& speeds) const override;
+  Twist2 twist(const WheelSpeeds &speeds) const override;
 
   /**
    * @brief      See \ref WheeledKinematics::wheel_speeds.
@@ -429,7 +548,7 @@ class NAVGROUND_CORE_EXPORT FourWheelsOmniDriveKinematics
    * @return     The corresponding wheel speeds in the order {front left, rear
    * left, rear right, rear left}
    */
-  WheelSpeeds wheel_speeds(const Twist2& twist) const override;
+  WheelSpeeds wheel_speeds(const Twist2 &twist) const override;
 
   /**
    * @brief      See \ref WheeledKinematics::feasible_wheel_speeds.
@@ -439,14 +558,14 @@ class NAVGROUND_CORE_EXPORT FourWheelsOmniDriveKinematics
    * @return     The corresponding wheel speeds in the order {front left, rear
    * left, rear right, rear left}
    */
-  WheelSpeeds feasible_wheel_speeds(const Twist2& twist) const override;
+  WheelSpeeds feasible_wheel_speeds(const Twist2 &twist) const override;
 
   /**
    * @private
    */
   std::string get_type() const override { return type; }
 
- private:
+private:
   const static std::string type;
 };
 
@@ -466,30 +585,39 @@ class NAVGROUND_CORE_EXPORT FourWheelsOmniDriveKinematics
  */
 class NAVGROUND_CORE_EXPORT DynamicTwoWheelsDifferentialDriveKinematics
     : public TwoWheelsDifferentialDriveKinematics {
- public:
+public:
   /**
    * @brief      Constructs a new instance.
    *
    * @param[in]  max_speed  The maximal wheel speed
    * @param[in]  axis  The wheel axis (i.e., the distance between the wheels)
+   * @param[in]  max_angular_speed  The maximal angular speed, that is also
+   * limited by the 2 * max_speed / axis, if the axis is positive.
+   * @param[in]  max_forward_speed  The maximal linear speed when moving
+   * forwards (set to negative or infinite to leave unconstrained)
+   * @param[in]  max_backward_speed The maximal linear speed when moving
+   * backwards (set to negative or infinite to leave unconstrained)
    * @param[in]  max_acceleration The maximal linear body acceleration
    * @param[in]  moi The scaled moment of inertial (``moi = I / (mass * axis^2 /
    * 8)``) Equal to one for an homogeneous disc of diameter ``axis``.
    */
-  DynamicTwoWheelsDifferentialDriveKinematics(ng_float_t max_speed = 0,
-                                              ng_float_t axis = 0,
-                                              ng_float_t max_acceleration = 0,
-                                              ng_float_t moi = 1)
-      : TwoWheelsDifferentialDriveKinematics(max_speed, axis),
-        max_acceleration(max_acceleration),
-        moi(moi){}
+  DynamicTwoWheelsDifferentialDriveKinematics(
+      ng_float_t max_speed = Kinematics::inf, ng_float_t axis = 0,
+      ng_float_t max_angular_speed = Kinematics::inf,
+      ng_float_t max_forward_speed = Kinematics::inf,
+      ng_float_t max_backward_speed = 0,
+      ng_float_t max_acceleration = Kinematics::inf, ng_float_t moi = 1)
+      : TwoWheelsDifferentialDriveKinematics(max_speed, axis, max_angular_speed,
+                                             max_forward_speed,
+                                             max_backward_speed),
+        max_acceleration(max_acceleration), moi(moi) {}
 
   using TwoWheelsDifferentialDriveKinematics::feasible;
 
   /**
    * @private
    */
-  Twist2 feasible(const Twist2& twist, const Twist2& current,
+  Twist2 feasible(const Twist2 &twist, const Twist2 &current,
                   ng_float_t time_step) const override;
 
   /**
@@ -510,7 +638,8 @@ class NAVGROUND_CORE_EXPORT DynamicTwoWheelsDifferentialDriveKinematics
    * @param[in]  value  A positive value
    */
   void set_moi(ng_float_t value) {
-    if (value > 0) moi = value;
+    if (value > 0)
+      moi = value;
   }
 
   /**
@@ -540,7 +669,8 @@ class NAVGROUND_CORE_EXPORT DynamicTwoWheelsDifferentialDriveKinematics
    * @param[in]  value  A positive value
    */
   void set_max_acceleration(ng_float_t value) {
-    if (value > 0) max_acceleration = value;
+    if (value > 0)
+      max_acceleration = value;
   }
 
   /**
@@ -557,7 +687,8 @@ class NAVGROUND_CORE_EXPORT DynamicTwoWheelsDifferentialDriveKinematics
   void set_max_angular_acceleration(ng_float_t value);
 
   /**
-   * @brief      Computes the wheel torques required to accelerate over a time step
+   * @brief      Computes the wheel torques required to accelerate over a time
+   * step
    *
    * @param[in]  value      The target value
    * @param[in]  current    The current value
@@ -565,29 +696,31 @@ class NAVGROUND_CORE_EXPORT DynamicTwoWheelsDifferentialDriveKinematics
    *
    * @return     {left, right} wheel torques. May not be feasible
    */
-  std::vector<ng_float_t> wheel_torques(const Twist2& value, const Twist2& current, ng_float_t time_step) const;
-  
+  std::vector<ng_float_t> wheel_torques(const Twist2 &value,
+                                        const Twist2 &current,
+                                        ng_float_t time_step) const;
+
   /**
    * @brief      Applies wheel torques to accelerate a twist over a time step
    *
    *             Does not check the motor torques are feasible.
-   *  
+   *
    * @param[in]  values     The motor torques
    * @param[in]  current    The current twist
    * @param[in]  time_step  The time step
    *
    * @return     The accelerated twist.
    */
-  Twist2 twist_from_wheel_torques(const std::vector<ng_float_t> & values, const Twist2& current, ng_float_t time_step) const;
+  Twist2 twist_from_wheel_torques(const std::vector<ng_float_t> &values,
+                                  const Twist2 &current,
+                                  ng_float_t time_step) const;
 
   /**
    * @brief      Gets the maximal [scaled] wheel torque.
    *
    * @return     The maximal wheel torque (in acceleration units)
    */
-  ng_float_t get_max_wheel_torque() const {
-    return max_acceleration;
-  }
+  ng_float_t get_max_wheel_torque() const { return max_acceleration; }
 
   /**
    * @private
@@ -597,20 +730,20 @@ class NAVGROUND_CORE_EXPORT DynamicTwoWheelsDifferentialDriveKinematics
   /**
    * @private
    */
-  const Properties& get_properties() const override { return properties; };
+  const Properties &get_properties() const override { return properties; };
 
   /**
    * @private
    */
   static const std::map<std::string, Property> properties;
 
- private:
+private:
   const static std::string type;
   ng_float_t max_acceleration;
   ng_float_t moi;
   // bool reduce_torques;
 };
 
-}  // namespace navground::core
+} // namespace navground::core
 
 #endif /* end of include guard: navground_KINEMATICS_H_ */